requirements.txt
asgiref==3.8.1
Django==5.1.4
django-cors-headers==4.6.0
django-filter==24.3
djangorestframework==3.15.2
loguru==0.7.3
psycopg==3.2.4
python-dotenv==1.0.1
sqlparse==0.5.3
typing_extensions==4.12.2
